cdk/requirements.txt
aws-cdk-lib==2.31.1
constructs>=10.0.0,<11.0.0